Write versions of files to out/versions
The idea is we'll use `make` for the rest of the process, and having
these files containing the current versions will mean we only have to
recompute things which have new versions.

Note that make just uses the last modified times on the files, so the
version numbers don't matter to make (but they do let this script work
out when it needs to touch the files).